requirements.txt
asgiref==3.6.0
asttokens==2.0.7
attrs==22.1.0
backcall==0.2.0
black==22.6.0
build==0.10.0
CacheControl==0.12.11
certifi==2022.12.7
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.3
colorama==0.4.6
crashtest==0.4.1
decorator==5.1.1
distlib==0.3.6
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dulwich==0.21.3
executing==0.9.1
factory-boy==3.2.1
Faker==15.3.4
filelock==3.12.0
html5lib==1.1
idna==3.4
importlib-metadata==6.6.0
iniconfig==1.1.1
installer==0.7.0
ipdb==0.13.9
ipython==8.4.0
jaraco.classes==3.2.3
jedi==0.18.1
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
matplotlib-inline==0.1.3
model-bakery==1.9.0
more-itertools==9.1.0
msgpack==1.0.5
mypy-extensions==0.4.3
packaging==22.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
pkginfo==1.9.6
platformdirs==2.5.2
pluggy==1.0.0
poetry==1.4.2
poetry-core==1.5.2
poetry-plugin-export==1.3.1
prompt-toolkit==3.0.30
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.12.0
PyJWT==2.6.0
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytest==7.2.0
pytest-django==4.5.2
pytest-testdox==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32-ctypes==0.2.0
rapidfuzz==2.15.1
requests==2.28.2
requests-toolbelt==0.10.1
shellingham==1.5.0.post1
six==1.16.0
sqlparse==0.4.4
stack-data==0.3.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.7
traitlets==5.3.0
trove-classifiers==2023.4.22
tzdata==2023.3
urllib3==1.26.15
virtualenv==20.21.1
wcwidth==0.2.5
webencodings==0.5.1
zipp==3.15.0
